application/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.18
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.3
django-appconf==1.0.2
django-compressor==1.6
djangorestframework==3.5.2
drf-nested-routers==0.11.1
gunicorn==19.3.0
kombu==3.0.33
pi==0.1.2
pytz==2015.7
redis==2.10.5
six==1.9.0
static3==0.7.0
